Extract shared admin utilities into module
