+++ admin.html
@@ -295,6 +295,6 @@
     <!-- Import Data Modal -->
     <input type="file" id="importFile" style="display: none" accept=".json">
 
-    <script src="admin.js" defer></script>
+    <script type="module" src="admin.js"></script>
 </body>
 </html>
